add search drivers, driver get methods

+++ src/com/llc/bumpr/sdk/models/SearchQuery.java
@@ -4,10 +4,10 @@ import com.llc.bumpr.sdk.lib.Coordinate;
 
 public class SearchQuery {
 	
-	private float top;
-	private float right;
-	private float bottom;
-	private float left;
+	private double top;
+	private double right;
+	private double bottom;
+	private double left;
 	private double minFee;
 	private int minSeats;
 	
@@ -30,18 +30,18 @@ public class SearchQuery {
 	public static class Builder<T extends SearchQuery> {
 		T item;
 		
-		private float top;
-		private float right;
-		private float bottom;
-		private float left;
+		private double top;
+		private double right;
+		private double bottom;
+		private double left;
 		private double minFee;
 		private int minSeats;
 		
 		public Builder(T item) { this.item = item; }
-		public Builder<T> setTop(float top) { item.top = top; return this; }
-		public Builder<T> setRight(float right) { item.right = right; return this; }
-		public Builder<T> setBottom(float bottom) { item.bottom = bottom; return this; }
-		public Builder<T> setLeft(float left) { item.left = left; return this; }
+		public Builder<T> setTop(double top) { item.top = top; return this; }
+		public Builder<T> setRight(double right) { item.right = right; return this; }
+		public Builder<T> setBottom(double bottom) { item.bottom = bottom; return this; }
+		public Builder<T> setLeft(double left) { item.left = left; return this; }
 		public Builder<T> setMinFee(double minFee) { item.minFee = minFee; return this; }
 		public Builder<T> setMinSeats(int minSeats) { item.minSeats = minSeats; return this; }
